Combined two messages from two cases.
This resolution will allow the translator to decide if there should be a space between the two sentences.
In Japanese, the period (full stop) is used to separate from the next sentence. In English, we need a space character after the period. The .pot file will need to be regenerated.